gh-49: Renames BufferView.h to LooperView.h to reflect new naming.
